# book-network/src/main/java/com/ka/bsn/book/BookResponse.java
package com.ka.bsn.book;

public record BookResponse(
        Long id,
        String title,
        String authorName,
        String isbn,
        String synopsis,
        String owner,
        double rate,
        byte[] cover,
        boolean archived,
        boolean shareable
) {
}